requirements.txt
networkx
numpy
matplotlib
pyyaml
